Update import.cpp
[-Werror=conversion]

+++ src/opengl/main/gui/import.cpp
@@ -495,18 +495,20 @@ void import::collapsing_header_generated_tiles() const
           for (const auto &tile : tiles)
           {
                ImGui::TableNextColumn();
-               const auto       texSize = m_loaded_image_texture.get_size();
-               const glm::ivec4 rect    = { static_cast<int>(tile.x() / tile_size_px * selections->tile_size_value),
-                                            static_cast<int>(tile.y() / tile_size_px * selections->tile_size_value),
-                                            static_cast<int>(selections->tile_size_value),
-                                            static_cast<int>(selections->tile_size_value) };
-               const ImVec2     uv0     = { static_cast<float>(rect.x) / texSize.x, static_cast<float>(rect.y) / texSize.y };
+               const auto             texSize                 = m_loaded_image_texture.get_size();
+               const glm::ivec4       rect                    = { static_cast<int>(tile.x() / tile_size_px * selections->tile_size_value),
+                                                                  static_cast<int>(tile.y() / tile_size_px * selections->tile_size_value),
+                                                                  static_cast<int>(selections->tile_size_value),
+                                                                  static_cast<int>(selections->tile_size_value) };
+               const ImVec2           uv0                     = { static_cast<float>(rect.x) / static_cast<float>(texSize.x),
+                                                                  static_cast<float>(rect.y) / static_cast<float>(texSize.y) };
 
-               const ImVec2     uv1 = { static_cast<float>(rect.x + rect.z) / texSize.x, static_cast<float>(rect.y + rect.w) / texSize.y };
-               const auto       the_end_tile_table_tile = PushPopID();
-               static constexpr float button_size       = 32.F;
+               const ImVec2           uv1                     = { static_cast<float>(rect.x + rect.z) / static_cast<float>(texSize.x),
+                                                                  static_cast<float>(rect.y + rect.w) / static_cast<float>(texSize.y) };
+               const auto             the_end_tile_table_tile = PushPopID();
+               static constexpr float button_size             = 32.F;
 
-               const auto             str               = fmt::format("tb{}", i++);
+               const auto             str                     = fmt::format("tb{}", i++);
                ImGui::ImageButton(
                  str.c_str(),
                  glengine::ConvertGliDtoImTextureId<ImTextureID>(m_loaded_image_texture.id()),
